Migrate Image and Dataset types from @/services to @/types
