+++ graphcap_studio/src/common/index.ts
@@ -1,4 +1,3 @@
 // SPDX-License-Identifier: Apache-2.0
-export * from "./ui";
-export * from "./providers";
+
 export * from "../utils/index";
