+++ graphcap_studio/src/features/gallery-viewer/image-carousel/hooks/useCarouselNavigation.ts
@@ -1,4 +1,4 @@
-import { Image } from "@/services/images";
+import type { Image } from "@/types";
 // SPDX-License-Identifier: Apache-2.0
 import { useCallback, useEffect, useState } from "react";
