+++ graphcap_studio/src/features/image-properties/hooks/useImageProperties.ts
@@ -1,5 +1,5 @@
 import { useEditorContext } from "@/features/editor/context/EditorContext";
-import { Image } from "@/services/images";
+import type { Image } from "@/types";
 // SPDX-License-Identifier: Apache-2.0
 import { useEffect, useState } from "react";
