+++ graphcap_studio/src/types/index.ts
@@ -1,4 +1,9 @@
+// SPDX-License-Identifier: Apache-2.0
+export * from "./api-response-types";
+export * from "./dataset-types";
 export * from "./generation-option-types";
+export * from "./image-processing-types";
+export * from "./image-types";
 export * from "./perspective-module-types";
 export * from "./perspective-types";
 export * from "./provider-config-types";
