+++ graphcap_studio/src/utils/download.ts
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-import { Image } from "@/services/images";
+import type { Image } from "@/types";
 
 /**
  * Downloads an image from the given URL
